@@ -24,79 +30,93 @@
   ],
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/angular/angularfire2.git"
+    "url": "git+https://github.com/angular/angularfire.git"
   },
   "author": "davideast",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/angular/angularfire2/issues"
+    "url": "https://github.com/angular/angularfire/issues"
   },
-  "homepage": "https://github.com/angular/angularfire2#readme",
+  "homepage": "https://github.com/angular/angularfire#readme",
   "dependencies": {
-    "@angular-devkit/architect": "<0.900 || ^0.900.0-0 || ^9.0.0-0",
-    "@angular-devkit/core": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular-devkit/schematics": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/common": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/compiler": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/core": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/platform-browser": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/platform-browser-dynamic": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/router": ">=6.0.0 <9 || ^9.0.0-0",
-    "firebase": ">= 5.5.7 <7",
-    "firebase-tools": "^6.10.0",
+    "@angular-devkit/architect": ">= 0.900 < 0.1200",
+    "@angular-devkit/core": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@angular-devkit/schematics": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@angular/common": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@angular/compiler": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@angular/core": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@angular/platform-browser": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@angular/platform-browser-dynamic": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@angular/router": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "firebase": "^7.0.0 || ^8.0.0",
+    "firebase-admin": "^8.10.0",
+    "firebase-functions": "^3.6.0",
+    "firebase-tools": "^8.0.0",
+    "fs-extra": "^8.0.1",
     "fuzzy": "^0.1.3",
+    "husky": "^4.2.5",
     "inquirer": "^6.2.2",
     "inquirer-autocomplete-prompt": "^1.0.1",
-    "rxjs": "^6.0.0",
-    "ws": "^3.3.2",
+    "open": "^7.0.3",
+    "rxfire": "^3.9.7 || ^4.0.0",
+    "rxjs": "^6.5.3",
+    "semver": "^7.1.3",
+    "tslib": "^1.10.0 || ^2.0.0",
+    "ws": "^7.2.1",
     "xhr2": "^0.1.4",
-    "zone.js": "^0.8.0 || ^0.9.0"
+    "zone.js": "0.10.2"
   },
   "optionalDependencies": {
-    "bufferutil": "~3.0.0",
-    "utf-8-validate": "~4.0.0"
+    "bufferutil": "^4.0.1",
+    "utf-8-validate": "^5.0.2"
   },
   "devDependencies": {
-    "@angular/animations": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/compiler-cli": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/platform-server": ">=6.0.0 <9 || ^9.0.0-0",
+    "@angular-devkit/build-angular": ">= 0.900 < 0.1200",
+    "@angular-devkit/build-ng-packagr": ">= 0.900 < 0.1200",
+    "@angular/animations": " ^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@angular/cli": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@angular/compiler-cli": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@angular/platform-server": "^9.0.0 || ^10.0.0 || ^11.0.0",
+    "@firebase/app-types": "^0.6.1",
+    "@types/fs-extra": "^7.0.0",
+    "@types/gzip-size": "^5.1.1",
     "@types/inquirer": "^0.0.44",
-    "@types/jasmine": "^2.5.36",
+    "@types/jasmine": "^3.3.13",
+    "@types/node": "^12.6.2 < 12.12.42",
     "@types/request": "0.0.30",
+    "@types/semver": "^7.1.0",
+    "codelyzer": "^6.0.0",
     "concurrently": "^2.2.0",
     "conventional-changelog-cli": "^1.2.0",
-    "es6-module-loader": "^0.17.10",
-    "es6-shim": "^0.35.0",
-    "fs-extra": "^4.0.0",
-    "gulp": "^3.9.0",
-    "gulp-jasmine": "^2.2.1",
-    "gulp-typescript": "^2.10.0",
-    "gzip-size": "^4.0.0",
-    "http-server": "^0.8.5",
-    "jasmine": "^2.4.1",
-    "jasmine-core": "^2.4.1",
-    "json": "^9.0.3",
-    "karma": "^2.0.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-firefox-launcher": "^1.1.0",
-    "karma-jasmine": "^0.3.6",
-    "karma-mocha-reporter": "^2.0.2",
-    "karma-systemjs": "^0.10.0",
+    "firebase-functions-test": "^0.2.2",
+    "gzip-size": "^5.1.1",
+    "jasmine": "^3.4.0",
+    "jasmine-core": "~3.5.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "ncp": "^2.0.0",
-    "parse5": "^1.3.2",
+    "ng-packagr": "^10.1.0",
     "pretty-size": "^2.0.0",
-    "protractor": "3.0.0",
+    "protractor": "~7.0.0",
     "reflect-metadata": "0.1.2",
+    "replace-in-file": "^5.0.2",
     "rimraf": "^2.5.4",
-    "rollup": "^0.64.1",
-    "rollup-plugin-node-resolve": "^3.3.0",
-    "rollup-watch": "^4.3.1",
-    "schematics-utilities": "^1.1.1",
+    "rollup": "^2.22.0",
+    "schematics-utilities": "^2.0.1",
     "shelljs": "^0.8.0",
     "systemjs": "^0.19.16",
     "systemjs-builder": "^0.15.7",
     "traceur": "0.0.96",
-    "typescript": ">=3.4.0 <3.5.0"
+    "ts-transformer-keys": "^0.4.2",
+    "tsconfig-paths": "^3.8.0",
+    "tsickle": "^0.39.1",
+    "tslint": "~6.1.0",
+    "ttypescript": "^1.5.12",
+    "typedoc": "^0.16.4",
+    "typescript": "~3.9.7"
   },
   "typings": "index.d.ts"
 }
