@@ -1,38 +1,77 @@
-import { Injectable, InjectionToken } from '@angular/core';
-import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree, Router } from '@angular/router';
+import { Inject, Injectable, NgZone, Optional } from '@angular/core';
+import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from '@angular/router';
 import { Observable, of, pipe, UnaryFunction } from 'rxjs';
-import { map, switchMap, take } from 'rxjs/operators'
-import { User, auth } from 'firebase/app';
-import { AngularFireAuth } from '@angular/fire/auth';
+import { map, observeOn, shareReplay, switchMap, take, tap } from 'rxjs/operators';
+import firebase from 'firebase/app';
+import {
+  ɵAngularFireSchedulers,
+  FirebaseOptions,
+  FirebaseAppConfig,
+  FIREBASE_OPTIONS,
+  FIREBASE_APP_NAME,
+  ɵfirebaseAppFactory,
+  ɵkeepUnstableUntilFirstFactory
+} from '@angular/fire';
 
 export type AuthPipeGenerator = (next: ActivatedRouteSnapshot, state: RouterStateSnapshot) => AuthPipe;
-export type AuthPipe = UnaryFunction<Observable<User|null>, Observable<boolean|any[]>>;
+export type AuthPipe = UnaryFunction<Observable<firebase.User|null>, Observable<boolean|any[]>>;
 
-@Injectable()
+export const loggedIn: AuthPipe = map(user => !!user);
+
+@Injectable({
+  providedIn: 'any'
+})
 export class AngularFireAuthGuard implements CanActivate {
 
-  constructor(private afAuth: AngularFireAuth, private router: Router) {}
+  authState: Observable<firebase.User|null>;
 
-  canActivate(next: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
-    const authPipeFactory: AuthPipeGenerator = next.data.authGuardPipe || (() => loggedIn);
-    return this.afAuth.user.pipe(
-        take(1),
-        authPipeFactory(next, state),
-        map(canActivate => typeof canActivate == "boolean" ? canActivate : this.router.createUrlTree(canActivate))
+  constructor(
+    @Inject(FIREBASE_OPTIONS) options: FirebaseOptions,
+    @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig: string|FirebaseAppConfig|null|undefined,
+    zone: NgZone,
+    private router: Router
+  ) {
+
+    const schedulers = new ɵAngularFireSchedulers(zone);
+    const keepUnstableUntilFirst = ɵkeepUnstableUntilFirstFactory(schedulers);
+
+    const auth = of(undefined).pipe(
+      observeOn(new ɵAngularFireSchedulers(zone).outsideAngular),
+      switchMap(() => zone.runOutsideAngular(() => import('firebase/auth'))),
+      map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
+      map(app => zone.runOutsideAngular(() => app.auth())),
+      shareReplay({ bufferSize: 1, refCount: false }),
+    );
+
+    this.authState = auth.pipe(
+      switchMap(auth => new Observable<firebase.User|null>(auth.onAuthStateChanged.bind(auth))),
+      keepUnstableUntilFirst
+    );
+  }
+
+  canActivate = (next: ActivatedRouteSnapshot, state: RouterStateSnapshot) => {
+    const authPipeFactory = next.data.authGuardPipe as AuthPipeGenerator || (() => loggedIn);
+    return this.authState.pipe(
+      take(1),
+      authPipeFactory(next, state),
+      map(can => typeof can === 'boolean' ? can : this.router.createUrlTree(can as any[]))
     );
   }
 
 }
 
-export const canActivate = (pipe: AuthPipe|AuthPipeGenerator) => ({
-    canActivate: [ AngularFireAuthGuard ], data: { authGuardPipe: pipe.name === "" ? pipe : () => pipe}
+export const canActivate = (pipe: AuthPipeGenerator) => ({
+    canActivate: [ AngularFireAuthGuard ], data: { authGuardPipe: pipe }
 });
 
-export const loggedIn: AuthPipe = map(user => !!user);
+
 export const isNotAnonymous: AuthPipe = map(user => !!user && !user.isAnonymous);
-export const idTokenResult = switchMap((user: User|null) => user ? user.getIdTokenResult() : of(null));
+export const idTokenResult = switchMap((user: firebase.User|null) => user ? user.getIdTokenResult() : of(null));
 export const emailVerified: AuthPipe = map(user => !!user && user.emailVerified);
 export const customClaims = pipe(idTokenResult, map(idTokenResult => idTokenResult ? idTokenResult.claims : []));
-export const hasCustomClaim = (claim:string) => pipe(customClaims, map(claims =>  claims.hasOwnProperty(claim)));
-export const redirectUnauthorizedTo = (redirect: any[]) => pipe(loggedIn, map(loggedIn => loggedIn || redirect));
-export const redirectLoggedInTo = (redirect: any[]) =>  pipe(loggedIn, map(loggedIn => loggedIn && redirect || true));
+export const hasCustomClaim: (claim: string) => AuthPipe =
+  (claim) => pipe(customClaims, map(claims =>  claims.hasOwnProperty(claim)));
+export const redirectUnauthorizedTo: (redirect: any[]) => AuthPipe =
+  (redirect) => pipe(loggedIn, map(loggedIn => loggedIn || redirect));
+export const redirectLoggedInTo: (redirect: any[]) => AuthPipe =
+  (redirect) => pipe(loggedIn, map(loggedIn => loggedIn && redirect || true));
