@@ -1,16 +1,15 @@
-import { Observable, from } from 'rxjs';
+import { from, Observable } from 'rxjs';
 import { fromCollectionRef } from '../observable/fromRef';
-import { map, filter, scan } from 'rxjs/operators';
-import { firestore } from 'firebase/app';
+import { filter, map, observeOn, scan } from 'rxjs/operators';
+import firebase from 'firebase/app';
 
-import { DocumentChangeType, CollectionReference, Query, DocumentReference, DocumentData, DocumentChangeAction } from '../interfaces';
+import { CollectionReference, DocumentChangeAction, DocumentChangeType, DocumentData, DocumentReference, Query } from '../interfaces';
 import { docChanges, sortedChanges } from './changes';
 import { AngularFirestoreDocument } from '../document/document';
 import { AngularFirestore } from '../firestore';
-import { runInZone } from '@angular/fire';
 
 export function validateEventsArray(events?: DocumentChangeType[]) {
-  if(!events || events!.length === 0) {
+  if (!events || events.length === 0) {
     events = ['added', 'removed', 'modified'];
   }
   return events;
@@ -39,7 +38,7 @@ export function validateEventsArray(events?: DocumentChangeType[]) {
  * // Subscribe to changes as snapshots. This provides you data updates as well as delta updates.
  * fakeStock.valueChanges().subscribe(value => console.log(value));
  */
-export class AngularFirestoreCollection<T=DocumentData> {
+export class AngularFirestoreCollection<T= DocumentData> {
   /**
    * The constructor takes in a CollectionReference and Query to provide wrapper methods
    * for data operations and data streaming.
@@ -48,7 +47,6 @@ export class AngularFirestoreCollection<T=DocumentData> {
    * when you update data it is not updating data to the window of your query unless
    * the data fits the criteria of the query. See the AssociatedRefence type for details
    * on this implication.
-   * @param ref
    */
   constructor(
     public readonly ref: CollectionReference,
@@ -59,31 +57,23 @@ export class AngularFirestoreCollection<T=DocumentData> {
    * Listen to the latest change in the stream. This method returns changes
    * as they occur and they are not sorted by query order. This allows you to construct
    * your own data structure.
-   * @param events
    */
   stateChanges(events?: DocumentChangeType[]): Observable<DocumentChangeAction<T>[]> {
-    if(!events || events.length === 0) {
-      return this.afs.scheduler.keepUnstableUntilFirst(
-        this.afs.scheduler.runOutsideAngular(
-          docChanges<T>(this.query)
-        )
+    if (!events || events.length === 0) {
+      return docChanges<T>(this.query, this.afs.schedulers.outsideAngular).pipe(
+        this.afs.keepUnstableUntilFirst
       );
     }
-    return this.afs.scheduler.keepUnstableUntilFirst(
-        this.afs.scheduler.runOutsideAngular(
-          docChanges<T>(this.query)
-        )
-      )
-      .pipe(
-        map(actions => actions.filter(change => events.indexOf(change.type) > -1)),
-        filter(changes =>  changes.length > 0)
-      );
+    return docChanges<T>(this.query, this.afs.schedulers.outsideAngular).pipe(
+      map(actions => actions.filter(change => events.indexOf(change.type) > -1)),
+      filter(changes =>  changes.length > 0),
+      this.afs.keepUnstableUntilFirst
+    );
   }
 
   /**
    * Create a stream of changes as they occur it time. This method is similar to stateChanges()
    * but it collects each event in an array over time.
-   * @param events
    */
   auditTrail(events?: DocumentChangeType[]): Observable<DocumentChangeAction<T>[]> {
     return this.stateChanges(events).pipe(scan((current, action) => [...current, ...action], []));
@@ -92,50 +82,48 @@ export class AngularFirestoreCollection<T=DocumentData> {
   /**
    * Create a stream of synchronized changes. This method keeps the local array in sorted
    * query order.
-   * @param events
    */
   snapshotChanges(events?: DocumentChangeType[]): Observable<DocumentChangeAction<T>[]> {
     const validatedEvents = validateEventsArray(events);
-    const sortedChanges$ = sortedChanges<T>(this.query, validatedEvents);
-    const scheduledSortedChanges$ = this.afs.scheduler.runOutsideAngular(sortedChanges$);
-    return this.afs.scheduler.keepUnstableUntilFirst(scheduledSortedChanges$);
+    const scheduledSortedChanges$ = sortedChanges<T>(this.query, validatedEvents, this.afs.schedulers.outsideAngular);
+    return scheduledSortedChanges$.pipe(
+      this.afs.keepUnstableUntilFirst
+    );
   }
 
   /**
    * Listen to all documents in the collection and its possible query as an Observable.
-   * 
+   *
    * If the `idField` option is provided, document IDs are included and mapped to the
    * provided `idField` property name.
-   * @param options
    */
-  valueChanges(): Observable<T[]>
-  valueChanges({}): Observable<T[]>
-  valueChanges<K extends string>(options: {idField: K}): Observable<(T & { [T in K]: string })[]>
+  valueChanges(): Observable<T[]>;
+  // tslint:disable-next-line:unified-signatures
+  valueChanges({}): Observable<T[]>;
+  valueChanges<K extends string>(options: {idField: K}): Observable<(T & { [T in K]: string })[]>;
   valueChanges<K extends string>(options: {idField?: K} = {}): Observable<T[]> {
-    const fromCollectionRef$ = fromCollectionRef<T>(this.query);
-    const scheduled$ = this.afs.scheduler.runOutsideAngular(fromCollectionRef$);
-    return this.afs.scheduler.keepUnstableUntilFirst(scheduled$)
+    return fromCollectionRef<T>(this.query, this.afs.schedulers.outsideAngular)
       .pipe(
         map(actions => actions.payload.docs.map(a => {
           if (options.idField) {
-            return { 
-              ...a.data() as Object, 
-              ...{ [options.idField]: a.id } 
+            return {
+              ...a.data() as {},
+              ...{ [options.idField]: a.id }
             } as T & { [T in K]: string };
           } else {
-            return a.data()
+            return a.data();
           }
-        }))
+        })),
+        this.afs.keepUnstableUntilFirst
       );
   }
 
   /**
-   * Retrieve the results of the query once. 
-   * @param options 
+   * Retrieve the results of the query once.
    */
-  get(options?: firestore.GetOptions) {
+  get(options?: firebase.firestore.GetOptions) {
     return from(this.query.get(options)).pipe(
-      runInZone(this.afs.scheduler.zone)
+      observeOn(this.afs.schedulers.insideAngular),
     );
   }
 
@@ -152,9 +140,8 @@ export class AngularFirestoreCollection<T=DocumentData> {
 
   /**
    * Create a reference to a single document in a collection.
-   * @param path
    */
-  doc<T>(path?: string): AngularFirestoreDocument<T> {
-    return new AngularFirestoreDocument<T>(this.ref.doc(path), this.afs);
+  doc<T2 = T>(path?: string): AngularFirestoreDocument<T2> {
+    return new AngularFirestoreDocument<T2>(this.ref.doc(path), this.afs);
   }
 }
