@@ -1,12 +1,10 @@
-import { Observable, from } from 'rxjs';
-import { DocumentReference, SetOptions, DocumentData, QueryFn, Action, DocumentSnapshot } from '../interfaces';
+import { from, Observable } from 'rxjs';
+import { Action, DocumentData, DocumentReference, DocumentSnapshot, QueryFn, SetOptions } from '../interfaces';
 import { fromDocRef } from '../observable/fromRef';
-import { map } from 'rxjs/operators';
-
+import { map, observeOn } from 'rxjs/operators';
 import { AngularFirestore, associateQuery } from '../firestore';
 import { AngularFirestoreCollection } from '../collection/collection';
-import { firestore } from 'firebase/app';
-import { runInZone } from '@angular/fire';
+import firebase from 'firebase/app';
 
 /**
  * AngularFirestoreDocument service
@@ -35,14 +33,11 @@ export class AngularFirestoreDocument<T = DocumentData> {
   /**
    * The contstuctor takes in a DocumentReference to provide wrapper methods
    * for data operations, data streaming, and Symbol.observable.
-   * @param ref
    */
   constructor(public ref: DocumentReference, private afs: AngularFirestore) { }
 
   /**
    * Create or overwrite a single document.
-   * @param data
-   * @param options
    */
   set(data: T, options?: SetOptions): Promise<void> {
     return this.ref.set(data, options);
@@ -50,7 +45,6 @@ export class AngularFirestoreDocument<T = DocumentData> {
 
   /**
    * Update some fields of a document without overwriting the entire document.
-   * @param data
    */
   update(data: Partial<T>): Promise<void> {
     return this.ref.update(data);
@@ -66,8 +60,6 @@ export class AngularFirestoreDocument<T = DocumentData> {
   /**
    * Create a reference to a sub-collection given a path and an optional query
    * function.
-   * @param path
-   * @param queryFn
    */
   collection<R = DocumentData>(path: string, queryFn?: QueryFn): AngularFirestoreCollection<R> {
     const collectionRef = this.ref.collection(path);
@@ -79,9 +71,10 @@ export class AngularFirestoreDocument<T = DocumentData> {
    * Listen to snapshot updates from the document.
    */
   snapshotChanges(): Observable<Action<DocumentSnapshot<T>>> {
-    const fromDocRef$ = fromDocRef<T>(this.ref);
-    const scheduledFromDocRef$ = this.afs.scheduler.runOutsideAngular(fromDocRef$);
-    return this.afs.scheduler.keepUnstableUntilFirst(scheduledFromDocRef$);
+    const scheduledFromDocRef$ = fromDocRef<T>(this.ref, this.afs.schedulers.outsideAngular);
+    return scheduledFromDocRef$.pipe(
+      this.afs.keepUnstableUntilFirst
+    );
   }
 
   /**
@@ -107,11 +100,10 @@ export class AngularFirestoreDocument<T = DocumentData> {
 
   /**
    * Retrieve the document once.
-   * @param options
    */
-  get(options?: firestore.GetOptions) {
+  get(options?: firebase.firestore.GetOptions) {
     return from(this.ref.get(options)).pipe(
-      runInZone(this.afs.scheduler.zone)
+      observeOn(this.afs.schedulers.insideAngular),
     );
   }
 }
