@@ -1,26 +1,49 @@
-import { Injectable, Inject, Optional, NgZone, PLATFORM_ID } from '@angular/core';
-import { DatabaseQuery, PathReference, DatabaseSnapshot, ChildEvent, ListenEvent, QueryFn, AngularFireList, AngularFireObject } from './interfaces';
+import { Inject, Injectable, InjectionToken, NgZone, Optional, PLATFORM_ID } from '@angular/core';
+import { AngularFireList, AngularFireObject, DatabaseQuery, PathReference, QueryFn } from './interfaces';
 import { getRef } from './utils';
-import { InjectionToken } from '@angular/core';
 import { createListReference } from './list/create-reference';
 import { createObjectReference } from './object/create-reference';
-import { FirebaseDatabase, FirebaseOptions, FirebaseAppConfig, FirebaseOptionsToken, FirebaseNameOrConfigToken, RealtimeDatabaseURL, _firebaseAppFactory, FirebaseZoneScheduler } from '@angular/fire';
+import {
+  FIREBASE_APP_NAME,
+  FIREBASE_OPTIONS,
+  FirebaseAppConfig,
+  FirebaseOptions,
+  ɵAngularFireSchedulers,
+  ɵfirebaseAppFactory,
+  ɵkeepUnstableUntilFirstFactory
+} from '@angular/fire';
+import { Observable } from 'rxjs';
+import 'firebase/database';
+import { registerDatabase } from '@firebase/database';
+import firebase from 'firebase/app';
 
-@Injectable()
+export const URL = new InjectionToken<string>('angularfire2.realtimeDatabaseURL');
+
+@Injectable({
+  providedIn: 'any'
+})
 export class AngularFireDatabase {
-  public readonly database: FirebaseDatabase;
-  public readonly scheduler: FirebaseZoneScheduler;
+  public readonly database: firebase.database.Database;
+
+  public readonly schedulers: ɵAngularFireSchedulers;
+  public readonly keepUnstableUntilFirst: <T>(obs$: Observable<T>) => Observable<T>;
 
   constructor(
-    @Inject(FirebaseOptionsToken) options:FirebaseOptions,
-    @Optional() @Inject(FirebaseNameOrConfigToken) nameOrConfig:string|FirebaseAppConfig|null|undefined,
-    @Optional() @Inject(RealtimeDatabaseURL) databaseURL:string|null,
+    @Inject(FIREBASE_OPTIONS) options: FirebaseOptions,
+    @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig: string | FirebaseAppConfig | null | undefined,
+    @Optional() @Inject(URL) databaseURL: string | null,
+    // tslint:disable-next-line:ban-types
     @Inject(PLATFORM_ID) platformId: Object,
     zone: NgZone
   ) {
-    this.scheduler = new FirebaseZoneScheduler(zone, platformId);
+    this.schedulers = new ɵAngularFireSchedulers(zone);
+    this.keepUnstableUntilFirst = ɵkeepUnstableUntilFirstFactory(this.schedulers);
+
     this.database = zone.runOutsideAngular(() => {
-      const app = _firebaseAppFactory(options, nameOrConfig);
+      const app = ɵfirebaseAppFactory(options, zone, nameOrConfig);
+      if (registerDatabase) {
+        registerDatabase(firebase as any);
+      }
       return app.database(databaseURL || undefined);
     });
   }
@@ -28,13 +51,13 @@ export class AngularFireDatabase {
   list<T>(pathOrRef: PathReference, queryFn?: QueryFn): AngularFireList<T> {
     const ref = getRef(this.database, pathOrRef);
     let query: DatabaseQuery = ref;
-    if(queryFn) {
+    if (queryFn) {
       query = queryFn(ref);
     }
     return createListReference<T>(query, this);
   }
 
-  object<T>(pathOrRef: PathReference): AngularFireObject<T>  {
+  object<T>(pathOrRef: PathReference): AngularFireObject<T> {
     const ref = getRef(this.database, pathOrRef);
     return createObjectReference<T>(ref, this);
   }
@@ -57,5 +80,3 @@ export {
   Action,
   SnapshotAction
 } from './interfaces';
-
-export { RealtimeDatabaseURL };
