@@ -1,4 +1,4 @@
-import { DatabaseQuery, AngularFireList, ChildEvent } from '../interfaces';
+import { AngularFireList, ChildEvent, DatabaseQuery } from '../interfaces';
 import { snapshotChanges } from './snapshot-changes';
 import { stateChanges } from './state-changes';
 import { auditTrail } from './audit-trail';
@@ -7,46 +7,30 @@ import { createRemoveMethod } from './remove';
 import { AngularFireDatabase } from '../database';
 import { map } from 'rxjs/operators';
 
-export function createListReference<T=any>(query: DatabaseQuery, afDatabase: AngularFireDatabase): AngularFireList<T> {
+export function createListReference<T= any>(query: DatabaseQuery, afDatabase: AngularFireDatabase): AngularFireList<T> {
+  const outsideAngularScheduler = afDatabase.schedulers.outsideAngular;
+  const refInZone = afDatabase.schedulers.ngZone.run(() => query.ref);
   return {
     query,
-    update: createDataOperationMethod<Partial<T>>(query.ref, 'update'),
-    set: createDataOperationMethod<T>(query.ref, 'set'),
-    push: (data: T) => query.ref.push(data),
-    remove: createRemoveMethod(query.ref),
+    update: createDataOperationMethod<Partial<T>>(refInZone, 'update'),
+    set: createDataOperationMethod<T>(refInZone, 'set'),
+    push: (data: T) => refInZone.push(data),
+    remove: createRemoveMethod(refInZone),
     snapshotChanges(events?: ChildEvent[]) {
-      const snapshotChanges$ = snapshotChanges<T>(query, events);
-      return afDatabase.scheduler.keepUnstableUntilFirst(
-        afDatabase.scheduler.runOutsideAngular(
-          snapshotChanges$
-        )
-      );
+      return snapshotChanges<T>(query, events, outsideAngularScheduler).pipe(afDatabase.keepUnstableUntilFirst);
     },
     stateChanges(events?: ChildEvent[]) {
-      const stateChanges$ = stateChanges<T>(query, events);
-      return afDatabase.scheduler.keepUnstableUntilFirst(
-        afDatabase.scheduler.runOutsideAngular(
-          stateChanges$
-        )
-      );
+      return stateChanges<T>(query, events, outsideAngularScheduler).pipe(afDatabase.keepUnstableUntilFirst);
     },
     auditTrail(events?: ChildEvent[]) {
-      const auditTrail$ = auditTrail<T>(query, events)
-      return afDatabase.scheduler.keepUnstableUntilFirst(
-        afDatabase.scheduler.runOutsideAngular(
-          auditTrail$
-        )
-      );
+      return auditTrail<T>(query, events, outsideAngularScheduler).pipe(afDatabase.keepUnstableUntilFirst);
     },
     valueChanges(events?: ChildEvent[]) {
-      const snapshotChanges$ = snapshotChanges<T>(query, events);
-      return afDatabase.scheduler.keepUnstableUntilFirst(
-        afDatabase.scheduler.runOutsideAngular(
-          snapshotChanges$
-        )
-      ).pipe(
-        map(actions => actions.map(a => a.payload.val() as T))
+      const snapshotChanges$ = snapshotChanges<T>(query, events, outsideAngularScheduler);
+      return snapshotChanges$.pipe(
+        map(actions => actions.map(a => a.payload.val() as T)),
+        afDatabase.keepUnstableUntilFirst
       );
     }
-  }
+  };
 }
