@@ -1,36 +1,37 @@
-import { FirebaseApp, AngularFireModule } from '@angular/fire';
-import { AngularFirestore } from '../firestore';
+import { AngularFireModule, FirebaseApp } from '@angular/fire';
+import { AngularFirestore, SETTINGS } from '../firestore';
 import { AngularFirestoreModule } from '../firestore.module';
-import { AngularFirestoreDocument } from '../document/document';
 import { Subscription } from 'rxjs';
+import { AngularFirestoreDocument } from './document';
 import { take } from 'rxjs/operators';
 
-import { TestBed, inject } from '@angular/core/testing';
-import { COMMON_CONFIG } from '../test-config';
+import { TestBed } from '@angular/core/testing';
+import { COMMON_CONFIG } from '../../test-config';
 
-import { Stock, randomName, FAKE_STOCK_DATA } from '../utils.spec';
+import { FAKE_STOCK_DATA, rando, randomName, Stock } from '../utils.spec';
+import 'firebase/firestore';
 
 describe('AngularFirestoreDocument', () => {
   let app: FirebaseApp;
   let afs: AngularFirestore;
-  let sub: Subscription;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG),
-        AngularFirestoreModule.enablePersistence({ synchronizeTabs: true })
+        AngularFireModule.initializeApp(COMMON_CONFIG, rando()),
+        AngularFirestoreModule
+      ],
+      providers: [
+        { provide: SETTINGS, useValue: { host: 'localhost:8080', ssl: false } }
       ]
     });
-    inject([FirebaseApp, AngularFirestore], (_app: FirebaseApp, _afs: AngularFirestore) => {
-      app = _app;
-      afs = _afs;
-    })();
+
+    app = TestBed.inject(FirebaseApp);
+    afs = TestBed.inject(AngularFirestore);
   });
 
-  afterEach(done => {
+  afterEach(() => {
     app.delete();
-    done();
   });
 
   describe('valueChanges()', () => {
@@ -79,6 +80,17 @@ describe('AngularFirestoreDocument', () => {
             stock.delete().then(done).catch(done.fail);
           }
         });
+    });
+    
+  it('should get unwrapped snapshot', async (done: any) => {
+    const randomCollectionName = afs.firestore.collection('a').doc().id;
+    const ref = afs.firestore.doc(`${randomCollectionName}/FAKE`);
+    const stock = new AngularFirestoreDocument<Stock>(ref, afs);
+    await stock.set(FAKE_STOCK_DATA);
+    const obs$ = stock.valueChanges();
+    obs$.pipe(take(1)).subscribe(async data => {
+      expect(data).toEqual(FAKE_STOCK_DATA);
+      stock.delete().then(done).catch(done.fail);
     });
 
   });
