@@ -1,41 +1,35 @@
-import { FirebaseApp, FirebaseOptionsToken, AngularFireModule, FirebaseNameOrConfigToken } from '@angular/fire';
-import { AngularFirestore } from './firestore';
+import { AngularFireModule, FIREBASE_APP_NAME, FIREBASE_OPTIONS, FirebaseApp } from '@angular/fire';
+import { AngularFirestore, SETTINGS } from './firestore';
 import { AngularFirestoreModule } from './firestore.module';
 import { AngularFirestoreDocument } from './document/document';
 import { AngularFirestoreCollection } from './collection/collection';
 
-import { Observable, Subscription } from 'rxjs';
-
-import { TestBed, inject } from '@angular/core/testing';
-import { COMMON_CONFIG } from './test-config';
-
-interface Stock {
-  name: string;
-  price: number;
-}
+import { TestBed } from '@angular/core/testing';
+import { COMMON_CONFIG } from '../test-config';
+import 'firebase/firestore';
+import { rando } from './utils.spec';
 
 describe('AngularFirestore', () => {
   let app: FirebaseApp;
   let afs: AngularFirestore;
-  let sub: Subscription;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG),
-        AngularFirestoreModule.enablePersistence({synchronizeTabs: true})
+        AngularFireModule.initializeApp(COMMON_CONFIG, rando()),
+        AngularFirestoreModule.enablePersistence()
+      ],
+      providers: [
+        { provide: SETTINGS, useValue: { host: 'localhost:8080', ssl: false } }
       ]
     });
-    inject([FirebaseApp, AngularFirestore], (_app: FirebaseApp, _afs: AngularFirestore) => {
-      app = _app;
-      afs = _afs;
-    })();
+
+    app = TestBed.inject(FirebaseApp);
+    afs = TestBed.inject(AngularFirestore);
   });
 
-  afterEach(done => {
-    // can't await here https://github.com/firebase/firebase-js-sdk/issues/605
+  afterEach(() => {
     app.delete();
-    done();
   });
 
   it('should be the properly initialized type', () => {
@@ -44,7 +38,6 @@ describe('AngularFirestore', () => {
 
   it('should have an initialized Firebase app', () => {
     expect(afs.firestore.app).toBeDefined();
-    expect(afs.firestore.app).toEqual(app);
   });
 
   it('should create an AngularFirestoreDocument from a string path', () => {
@@ -53,8 +46,7 @@ describe('AngularFirestore', () => {
   });
 
   it('should create an AngularFirestoreDocument from a string path', () => {
-    const ref = afs.doc('a/doc').ref;
-    const doc = afs.doc(ref);
+    const doc = afs.doc(afs.doc('a/doc').ref);
     expect(doc instanceof AngularFirestoreDocument).toBe(true);
   });
 
@@ -64,8 +56,7 @@ describe('AngularFirestore', () => {
   });
 
   it('should create an AngularFirestoreCollection from a reference', () => {
-    const ref = afs.collection('stuffs').ref;
-    const collection = afs.collection(ref);
+    const collection = afs.collection(afs.collection('stuffs').ref);
     expect(collection instanceof AngularFirestoreCollection).toBe(true);
   });
 
@@ -83,41 +74,54 @@ describe('AngularFirestore', () => {
     expect(quadWrapper).toThrowError();
   });
 
-  it('should enable persistence', (done) => {
-    const sub = afs.persistenceEnabled$.subscribe(isEnabled => {
-      expect(isEnabled).toBe(true);
-      done();
+  if (typeof window === 'undefined') {
+
+    it('should not enable persistence (Node.js)', (done) => {
+      afs.persistenceEnabled$.subscribe(isEnabled => {
+        expect(isEnabled).toBe(false);
+        done();
+      });
     });
-  });
+
+  } else {
+
+    it('should enable persistence', (done) => {
+      afs.persistenceEnabled$.subscribe(isEnabled => {
+        expect(isEnabled).toBe(true);
+        done();
+      });
+    });
+
+  }
 
 });
-
-const FIREBASE_APP_NAME_TOO = (Math.random() + 1).toString(36).substring(7);
 
 describe('AngularFirestore with different app', () => {
   let app: FirebaseApp;
   let afs: AngularFirestore;
+  let firebaseAppName: string;
 
   beforeEach(() => {
+    firebaseAppName = rando();
     TestBed.configureTestingModule({
       imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG),
+        AngularFireModule.initializeApp(COMMON_CONFIG, rando()),
         AngularFirestoreModule
       ],
       providers: [
-        { provide: FirebaseNameOrConfigToken, useValue: FIREBASE_APP_NAME_TOO },
-        { provide: FirebaseOptionsToken, useValue: COMMON_CONFIG }
+        { provide: FIREBASE_APP_NAME, useValue: firebaseAppName },
+        { provide: FIREBASE_OPTIONS, useValue: COMMON_CONFIG },
+        { provide: SETTINGS, useValue: { host: 'localhost:8080', ssl: false } }
       ]
     });
-    inject([FirebaseApp, AngularFirestore], (app_: FirebaseApp, _afs: AngularFirestore) => {
-      app = app_;
-      afs = _afs;
-    })();
+
+
+    app = TestBed.inject(FirebaseApp);
+    afs = TestBed.inject(AngularFirestore);
   });
 
-  afterEach(done => {
+  afterEach(() => {
     app.delete();
-    done();
   });
 
   describe('<constructor>', () => {
@@ -128,11 +132,10 @@ describe('AngularFirestore with different app', () => {
 
     it('should have an initialized Firebase app', () => {
       expect(afs.firestore.app).toBeDefined();
-      expect(afs.firestore.app).toEqual(app);
     });
 
     it('should have an initialized Firebase app instance member', () => {
-      expect(afs.firestore.app.name).toEqual(FIREBASE_APP_NAME_TOO);
+      expect(afs.firestore.app.name).toEqual(firebaseAppName);
     });
   });
 
@@ -146,19 +149,20 @@ describe('AngularFirestore without persistance', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG),
+        AngularFireModule.initializeApp(COMMON_CONFIG, rando()),
         AngularFirestoreModule
+      ],
+      providers: [
+        { provide: SETTINGS, useValue: { host: 'localhost:8080', ssl: false } }
       ]
     });
-    inject([FirebaseApp, AngularFirestore], (_app: FirebaseApp, _afs: AngularFirestore) => {
-      app = _app;
-      afs = _afs;
-    })();
+
+    app = TestBed.inject(FirebaseApp);
+    afs = TestBed.inject(AngularFirestore);
   });
 
-  afterEach(done => {
+  afterEach(() => {
     app.delete();
-    done();
   });
 
   it('should not enable persistence', (done) => {
