@@ -1,19 +1,18 @@
-import { readFileSync } from "fs";
-import { FirebaseRc, Project } from "./interfaces";
-import { join } from "path";
+import { readFileSync } from 'fs';
+import { FirebaseRc, Project, WorkspaceProject } from './interfaces';
+import { join } from 'path';
+import { isUniversalApp } from './ng-add-ssr';
 
-export function listProjects() {
+export async function listProjects() {
   const firebase = require('firebase-tools');
-  return firebase.list().catch(
-    /* If list failed, then login and try again. */
-    () => firebase.login().then(() => firebase.list())
-  );
+  await firebase.login();
+  return firebase.projects.list();
 }
 
 // `fuzzy` passes either the original list of projects or an internal object
 // which contains the project as a property.
 const isProject = (elem: Project | { original: Project }): elem is Project => {
-  return (<{ original: Project }>elem).original === undefined;
+  return (elem as { original: Project }).original === undefined;
 };
 
 const searchProjects = (projects: Project[]) => {
@@ -22,7 +21,7 @@ const searchProjects = (projects: Project[]) => {
       require('fuzzy')
         .filter(input, projects, {
           extract(el: Project) {
-            return `${el.id} ${el.name} ${el.permission}`;
+            return `${el.projectId} ${el.displayName}`;
           }
         })
         .map((result: Project | { original: Project }) => {
@@ -33,9 +32,9 @@ const searchProjects = (projects: Project[]) => {
             original = result.original;
           }
           return {
-            name: `${original.id} (${original.name})`,
-            title: original.name,
-            value: original.id
+            name: `${original.displayName} (${original.projectId})`,
+            title: original.displayName,
+            value: original.projectId
           };
         })
     );
@@ -45,26 +44,38 @@ const searchProjects = (projects: Project[]) => {
 export const projectPrompt = (projects: Project[]) => {
   const inquirer = require('inquirer');
   inquirer.registerPrompt(
-    "autocomplete",
-    require("inquirer-autocomplete-prompt")
+    'autocomplete',
+    require('inquirer-autocomplete-prompt')
   );
   return inquirer.prompt({
-    type: "autocomplete",
-    name: "firebaseProject",
+    type: 'autocomplete',
+    name: 'firebaseProject',
     source: searchProjects(projects),
-    message: "Please select a project:"
+    message: 'Please select a project:'
   });
 };
 
+export const projectTypePrompt = (project: WorkspaceProject) => {
+  if (isUniversalApp(project)) {
+    return require('inquirer').prompt({
+      type: 'confirm',
+      name: 'universalProject',
+      message: 'We detected an Angular Universal project. Do you want to deploy as a Firebase Function?'
+    });
+  }
+  return Promise.resolve({ universalProject: false });
+};
+
 export function getFirebaseProjectName(
-  projectRoot: string,
+  workspaceRoot: string,
   target: string
 ): string | undefined {
-  const { targets }: FirebaseRc = JSON.parse(
-    readFileSync(join(projectRoot, ".firebaserc"), "UTF-8")
+  const rc: FirebaseRc = JSON.parse(
+    readFileSync(join(workspaceRoot, '.firebaserc'), 'UTF-8')
   );
-  const projects = Object.keys(targets!);
+  const targets = rc.targets || {};
+  const projects = Object.keys(targets || {});
   return projects.find(
-    project => !!Object.keys(targets![project].hosting).find(t => t === target)
+    project => !!Object.keys(targets[project].hosting).find(t => t === target)
   );
 }
